quote api work for but need designId for testing

+++ src/main/java/com/github/gcauchis/scalablepress/json/Sides.java
@@ -1,3 +1,25 @@
+/**
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2016 Gabriel Cauchis
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to
+ * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+ * the Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+ * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
 package com.github.gcauchis.scalablepress.json;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
@@ -14,52 +36,52 @@ public class Sides {
      * DTG: set to any positive number to indicate printing on this side,
      * Screenprint: the number of colors used on this side
      */
-    private Float front;
+    private Integer front;
     /**
      * DTG: set to any positive number to indicate printing on this side,
      * Screenprint: the number of colors used on this side
      */
-    private Float back;
+    private Integer back;
     /**
      * DTG: set to any positive number to indicate printing on this side,
      * Screenprint: the number of colors used on this side
      */
-    private Float right;
+    private Integer right;
     /**
      * DTG: set to any positive number to indicate printing on this side,
      * Screenprint: the number of colors used on this side
      */
-    private Float left;
+    private Integer left;
 
-    public Float getFront() {
+    public Integer getFront() {
         return front;
     }
 
-    public void setFront(Float front) {
+    public void setFront(Integer front) {
         this.front = front;
     }
 
-    public Float getBack() {
+    public Integer getBack() {
         return back;
     }
 
-    public void setBack(Float back) {
+    public void setBack(Integer back) {
         this.back = back;
     }
 
-    public Float getRight() {
+    public Integer getRight() {
         return right;
     }
 
-    public void setRight(Float right) {
+    public void setRight(Integer right) {
         this.right = right;
     }
 
-    public Float getLeft() {
+    public Integer getLeft() {
         return left;
     }
 
-    public void setLeft(Float left) {
+    public void setLeft(Integer left) {
         this.left = left;
     }
